@@ -22,36 +22,38 @@
   "dependencies": {
     "classnames": "^2.2.5",
     "codemirror": "^5.31.0",
+    "connected-react-router": "^6.5.0",
+    "form-data": "^2.3.2",
     "history": "^4.7.2",
     "immutable": "^3.8.2",
     "linq": "^3.0.9",
-    "marked": "^0.3.6",
+    "marked": "^0.6.2",
     "moment": "^2.19.1",
     "prop-types": "^15.6.0",
+    "query-string": "^6.8.1",
     "react-dock": "^0.2.4",
     "react-helmet": "^5.2.0",
-    "react-highlight": "^0.10.0",
-    "react-redux": "^5.0.6",
-    "react-redux-toastr": "^7.1.6",
-    "react-router": "^4.2.0",
-    "react-router-dom": "^4.2.2",
-    "react-router-redux": "^5.0.0-alpha.8",
-    "redux": "^3.7.2",
-    "redux-thunk": "^2.2.0",
-    "semantic-ui-react": "^0.76.0",
-    "styled-components": "^4.0.2",
+    "react-highlight": "^0.12.0",
+    "react-redux": "^7.1.0",
+    "react-redux-toastr": "^7.5.1",
+    "react-router": "^5.0.1",
+    "react-router-dom": "^5.0.1",
+    "redux": "^4.0.1",
+    "redux-thunk": "^2.3.0",
+    "semantic-ui-react": "^0.87.2",
+    "styled-components": "^4.3.2",
     "swagger-converter": "^1.4.1",
-    "urijs": "^1.19.0",
-    "form-data": "^2.3.2"
+    "swagger-parser": "^6.0.2",
+    "urijs": "^1.19.0"
   },
   "peerDependencies": {
-    "react": "15.x"
+    "react": "16.x"
   },
   "devDependencies": {
     "nodemon": "^1.12.1",
-    "nwb": "0.19.x",
-    "react": "^16.0.0",
-    "react-dom": "^16.0.0"
+    "nwb": "0.23.x",
+    "react": "^16.3.0",
+    "react-dom": "^16.3.0"
   },
   "author": "",
   "homepage": "",
