@@ -13,11 +13,11 @@
   },
   "dependencies": {
     "apiexplorer": "^0.6.0",
-    "react": "^16.2.0",
-    "react-dom": "^16.2.0"
+    "react": "^16.3.0",
+    "react-dom": "^16.3.0"
   },
   "devDependencies": {
-    "nwb": "0.19.x"
+    "nwb": "0.23.x"
   },
   "author": "",
   "license": "MIT",
