@@ -2,6 +2,8 @@ module.exports = {
   type: 'react-component',
   npm: {
     esModules: true,
-    umd: false
+    umd: {
+      global: 'ApiExplorer'
+    }
   }
 }
