@@ -17,7 +17,7 @@ export default class LinkGenerator {
    * @return String
    */
   makeHref (api, operationSpec, params) {
-    const {host, basePath} = api
+    const { host, basePath } = api
     const scheme = api.schemes ? api.schemes[0] : 'http'
     const path = operationSpec.url
     let href = `${scheme}://${host}${basePath}${path}`
