@@ -16,7 +16,7 @@ class NotFound extends Component {
         <br />
         <AppLogo width={128} />
         <br /><br />
-        <span style={{fontSize: '40px'}}>This is not the page you are looking for..!.</span>
+        <span style={{ fontSize: '40px' }}>This is not the page you are looking for..!.</span>
         <br /><br />
         <a href='/'><button className='btn btn-success'>
           <span className='glyphicon glyphicon-home' /> Back Home
