@@ -53,7 +53,8 @@ class Welcome extends Component {
     return pathsByTags.select(g => {
       let key = g.key()
       const api = this.props.apis.get('byName').get(apiInfo.apiname)
-      const tagDescription = api.hasOwnProperty('tags') ? api.tags.find(t => t.name === key).description : key
+      const tag = api.hasOwnProperty('tags') ? api.tags.find(t => t.name === key) : undefined
+      const tagDescription = tag ? tag.description || tag.key : key
 
       return <Segment key={key}>
         <h4>{tagDescription}</h4>
