@@ -1,32 +1,31 @@
 import SwaggerParser from 'swagger-parser'
 
-export function swagger2Loader (config, { onLoadProgress, onNewAPI, onNewOperation, onNewDefinition, onLoadCompleted, onLoadError }) {
+export function openAPI3Loader (config, { onLoadProgress, onNewAPI, onNewOperation, onNewDefinition, onLoadCompleted, onLoadError }) {
   const url = config.url.getUrl()
   onLoadProgress(`Loading API Spec from ${url}`)
 
   return SwaggerParser.validate(url)
-    .then(function (api) {
+    .then(api => {
       let newApi = api
       let defaultHost = `${window.location.origin}/${config.basePath}`
-
       newApi = config.interceptor({ friendlyName: config.friendlyName, url: config.url }, api)
-      swagger2SpecLoader(newApi, config.friendlyName, config.slug, defaultHost, { onLoadProgress, onNewAPI, onNewOperation, onLoadCompleted, onLoadError })
+      openAPI3SpecLoader(newApi, config.friendlyName, config.slug, defaultHost, { onLoadProgress, onNewAPI, onNewOperation, onLoadCompleted, onLoadError })
     })
     .catch(function (err) {
       onLoadError(err)
     })
 }
 
-export function swagger2SpecLoader (apiSpec, friendlyName, slug, defaultHost, { onLoadProgress, onNewAPI, onNewOperation, onLoadCompleted, onLoadError }) {
+export function openAPI3SpecLoader (apiSpec, friendlyName, slug, defaultHost, { onLoadProgress, onNewAPI, onNewOperation, onLoadCompleted, onLoadError }) {
   onLoadProgress(`API Spec received with success`)
   onLoadProgress(`Starting API parsing`)
 
   // defaults
   apiSpec = Object.assign({
     definitions: [],
-    basePath: '/'
+    basePath: '/',
+    host: apiSpec.servers ? apiSpec.servers[0].url : defaultHost
   }, apiSpec)
-  apiSpec.host = `${apiSpec.schemes[0]}://${apiSpec.host}` || defaultHost
 
   onNewAPI(apiSpec)
 
@@ -45,6 +44,16 @@ export function swagger2SpecLoader (apiSpec, friendlyName, slug, defaultHost, { 
           if (!operation.tags || operation.tags.length === 0) {
             operation.tags = ['']
           }
+
+          if (operation.responses) {
+            Object.keys(operation.responses)
+              .forEach(response => {
+                if (operation.responses[response].content && operation.responses[response].content['application/json']) {
+                  operation.responses[response].schema = operation.responses[response].content['application/json'].schema
+                }
+              })
+          }
+
           const operationSpec = { id, url, httpMethod, ...operation }
           onLoadProgress(`Processing operation ${id}`)
           onNewOperation(operationSpec)
