@@ -1,7 +1,7 @@
 import './index.css'
 
 import React from 'react'
-import {render} from 'react-dom'
+import { render } from 'react-dom'
 
 import App from './App'
 
